### 2025-08-09 a0f8ce4
CompatHelper: bump compat for KrylovKit to 0.10, (keep existing compat)
