+++ Project.toml
@@ -45,7 +45,7 @@ DocumenterTools = "0.1.20"
 FFTW = "1"
 FillArrays = "1.13.0"
 JLD2 = "0.4, 0.5"
-KrylovKit = "0.9.5"
+KrylovKit = "0.9.5, 0.10"
 LazyGrids = "1.1.0"
 LinearAlgebra = "1.6"
 LinearMaps = "3.11.3"
