### 2025-10-16 29323fc
CompatHelper: bump compat for JLD2 to 0.6, (keep existing compat)
