+++ Project.toml
@@ -44,7 +44,7 @@ DispatchDoctor = "0.4.22"
 DocumenterTools = "0.1.20"
 FFTW = "1"
 FillArrays = "1.13.0"
-JLD2 = "0.4, 0.5"
+JLD2 = "0.4, 0.5, 0.6"
 KrylovKit = "0.9.5"
 LazyGrids = "1.1.0"
 LinearAlgebra = "1.6"
