### 2025-10-19 25ce8a5
CompatHelper: bump compat for JLD2 to 0.6, (keep existing compat)
