### 2025-10-21 #8771 @jatkinson1000
Hi there @giordano, @navidcy, @cwentland0, @smillerc, @anand-me!

This paper has been submitted to JOSS, the Journal of Open Source Software.
It is a Julia package for performing stability analyses of rotating stratified flows.

Would you be willing to review it?
Please indicate by replying to this thread.

You can find information about reviewing for JOSS [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html).
