### 2025-11-06 #8771 @alexfikl
@jatkinson1000 Following up on https://github.com/openjournals/joss-reviews/issues/8894#issuecomment-3496248065, the topic is not quite my area of expertise, but I'm happy to review, if needed.
