### 2025-11-06 #9368 @editorialbot
[REVIEW]: BiGSTARS.jl: A Julia package for bi-global stability analysis for rotating stratified flows
**Submitting author:** @subhk (<a href="http://orcid.org/0000-0001-9737-3345">Subhajit Kar</a>)
**Repository:** https://github.com/subhk/BiGSTARS.jl
**Branch with paper.md** (empty if default branch):
**Version:** 2.0.8
