### 2025-11-07 bdea2a9
CompatHelper: bump compat for KrylovKit to 0.10, (keep existing compat)
