### 2025-11-08 subhk/BiGSTARS.jl#14 @alexfikl
Support for Julia 1.12
The README says that the package only works on Julia 1.11. I just tried it on Julia 1.12 (on Arch Linux) and it seems like the tests pass just fine (with an update to `Manifest.toml`). Is there anything else that would be needed to have advertise support?

xref: https://github.com/openjournals/joss-reviews/issues/9368
