### 2025-11-08 subhk/BiGSTARS.jl#15 @alexfikl
Contributing instructions
Beyond installation instructions in the README, this seems to lack some brief instructions for running the tests and contributing to the project (i.e. something like https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors)

xref: https://github.com/openjournals/joss-reviews/issues/9368
