### 2025-11-11 b451731
Add JOSS paper status badge to README
Added a status badge for JOSS paper.
