### 2025-11-11 6a2d448
Fix badge status duplication in README.md
