### 2025-11-11 8a7435e
docs: fix typos and math formatting in Stone1971
