+++ examples/Stone1971.jl
@@ -3,18 +3,18 @@
 # ## Introduction
 # Baroclinic instability (BCI) arises when a rotating, stratified fluid has tilted density surfaces, 
 # enabling eddies to tap available potential energy and convert it to kinetic energy.
-# Stone (1971) [eady1949long](@cite) investigated non-hydrostatic effects on BCI using Eady’s framework. 
-# He found that as the $Ri$ decreases, the wavelength of the most unstable mode increases 
+# Stone (1971) [stone1971](@cite) investigated non-hydrostatic effects on BCI using Eady’s framework.
+# He found that as the Richardson number decreases, the wavelength of the most unstable mode increases
 # while the growth rate diminishes relative to predictions from the quasigeostrophic (QG) approximation.
 # 
 # The basic state is given by
 # ```math
 # \begin{align}
-#     B(y, z) &= Ri z - y, \\
+#     B(y, z) &= \text{Ri}\, z - y, \\
 #     U(y, z) &= z - {1}/{2},
 # \end{align}
 # ```
-# where ``Ri`` is the Richardson number. We aim to analyze the stability of the 
+# where ``\text{Ri}`` is the Richardson number. We aim to analyze the stability of the
 # above basic state against small perturbations. The perturbation variables are
 # defined as
 # ```math
@@ -35,15 +35,16 @@
 #     \frac{D \mathbf{u}}{Dt}
 #     + \Big(v \frac{\partial U}{\partial y} + w \frac{\partial U}{\partial z} \Big) \hat{x}
 #     + \hat{z} \times \mathbf{u} &=
-#     -\nabla p + \frac{1}{\epsilon} b \hat{z} + E \nabla^2 \mathbf{u}, 
+#     -\nabla p + \frac{1}{\epsilon} b \hat{z} + \text{E} \nabla^2 \mathbf{u},
 # \\
 #     \frac{Db}{Dt}
-#     +  v \frac{\partial B}{\partial y} + w \frac{\partial B}{\partial z} &= \frac{E}{Pr} \nabla^2 b, 
+#     +  v \frac{\partial B}{\partial y} + w \frac{\partial B}{\partial z} &=
+#     \frac{\text{E}}{\text{Pr}} \nabla^2 b,
 # \\
 #     \nabla \cdot \mathbf{u} &= 0,
 # \end{align}
 # ```
-# where 
+# where ``\text{E}`` is the Eckman number, ``\text{Pr}`` is the Prandtl number, and
 # ```math
 # \begin{align}
 #   D/Dt \equiv \partial/\partial t + U (\partial/\partial x)
@@ -67,26 +68,26 @@
 # and the buoyancy $b$ 
 # ```math
 # \begin{align}
-#     \frac{D}{Dt}\nabla^2 {w} 
-#     + \frac{1}{\epsilon^2} \frac{\partial \zeta}{\partial z} 
-#     &= \frac{1}{\epsilon^2} \nabla_h^2 b + E \nabla^4 w,
+#     \frac{D}{Dt}\nabla^2 {w}
+#     + \frac{1}{\epsilon^2} \frac{\partial \zeta}{\partial z}
+#     &= \frac{1}{\epsilon^2} \nabla_h^2 b + \text{E} \nabla^4 w,
 # \\
 #     \frac{D \zeta}{Dt}
 #     - \frac{\partial U}{\partial z}\frac{\partial w}{\partial y}
-#     - \frac{\partial w}{\partial z} &= E \nabla^2 \zeta, 
+#     - \frac{\partial w}{\partial z} &= \text{E} \nabla^2 \zeta,
 # \\
[…]
