### 2025-11-11 d30b47e
docs: fix typos and math formatting in Eady1949
