+++ docs/src/references.bib
@@ -52,5 +52,9 @@
   publisher={Taylor \& Francis}
 }
 
-
-
+@book{pedlosky2013geophysical,
+  title={Geophysical fluid dynamics},
+  author={Pedlosky, Joseph},
+  year={2013},
+  publisher={Springer Science \& Business Media}
+}
