+++ Project.toml
@@ -13,17 +13,13 @@ DispatchDoctor = "8d63f2c5-f18a-4cf2-ba9d-b3f60fc568c8"
 DocumenterTools = "35a29f4d-8980-5a13-9543-d66fff28ecb8"
 FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
 FillArrays = "1a297f60-69ca-5386-bcde-b61e274b549b"
-JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
 KrylovKit = "0b1a1467-8014-51b9-945f-bf0ae24f4b77"
 LazyGrids = "7031d0ef-c40d-4431-b2f8-61a8d2f650db"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 LinearMaps = "7a12625a-238d-50fd-b39a-03d52299707e"
-ModelingToolkit = "961ee093-0014-501f-94e3-6117800e7a78"
-NonlinearSolve = "8913a72c-1f9b-4ce2-8d82-65094dcecaec"
 Parameters = "d96e819e-fc66-5662-9728-84c9c7592b0a"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
-Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 SparseMatrixDicts = "5cb6c4b0-9b79-11e8-24c9-f9621d252589"
 SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
@@ -44,16 +40,12 @@ DispatchDoctor = "0.4.22"
 DocumenterTools = "0.1.20"
 FFTW = "1"
 FillArrays = "1.13.0"
-JLD2 = "0.4, 0.5"
 KrylovKit = "0.9.5"
 LazyGrids = "1.1.0"
 LinearAlgebra = "1.6"
 LinearMaps = "3.11.3"
-ModelingToolkit = "9.79.1, 10"
-NonlinearSolve = "4.8.5"
 Parameters = "0.12.1"
 Random = "1.6"
-Reexport = "0.2, 1"
 SparseMatrixDicts = "0.3.0"
 SpecialFunctions = "2.5.1"
 StaticArrays = "0.12, 1"
@@ -62,7 +54,7 @@ ToeplitzMatrices = "0.8.4"
 Trapz = "2.0.3"
 Tullio = "0.3.6"
 VectorInterface = "0.5"
-julia = "1.10, 1.11"
+julia = "1.10, 1.11, 1.12"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
