+++ Project.toml
@@ -20,6 +20,7 @@ LinearMaps = "7a12625a-238d-50fd-b39a-03d52299707e"
 Parameters = "d96e819e-fc66-5662-9728-84c9c7592b0a"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
 Random = "9a3f8284-a2c9-5f02-9a11-845980a1fd5c"
+Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 SparseMatrixDicts = "5cb6c4b0-9b79-11e8-24c9-f9621d252589"
 SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
@@ -46,6 +47,7 @@ LinearAlgebra = "1.6"
 LinearMaps = "3.11.3"
 Parameters = "0.12.1"
 Random = "1.6"
+Reexport = "0.2, 1"
 SparseMatrixDicts = "0.3.0"
 SpecialFunctions = "2.5.1"
 StaticArrays = "0.12, 1"
