### 2025-11-13 9a0afd1
Add support for Julia 1.12 in README
Updated the supported Julia versions in the README.
