+++ docs/src/installation_instructions.md
@@ -27,6 +27,7 @@ We're more than happy to help you get your model up and running again.
     The latest version of `BiGSTARS.jl` requires **at least Julia v1.6** to run.  
     Installing `BiGSTARS.jl` with an older version of Julia will instead install the latest version that is compatible with your Julia installation.
 
-    BiGSTARS.jl is continuously tested on **Julia v1.10** (the current long-term support) and **v1.11**.  
+    BiGSTARS.jl is continuously tested on **Julia v1.10** (the current long-term support), **v1.11** and 
+    **v1.12**.  
     _For optimal stability and reproducibility, we strongly recommend using one of these tested Julia versions._
 
