### 2025-11-14 978d252
CompatHelper: bump compat for KrylovKit to 0.10, (keep existing compat)
