+++ CONTRIBUTING.md
@@ -1,1 +1,14 @@
+# Contributors' Guide
+
+Thanks for your interest in contributing to BiGSTARS.jl!
+
+Contributions of all kinds are welcome: bug reports, documentation improvements, 
+new examples, performance tweaks, and core features.
+
+
+If you’re adding a new feature:
+ - Put core logic in src/ (not in examples/).
+ - Add tests in test/.
+ - If relevant, add or update an example in examples/.
+ - Update docs in docs/ or docstrings.
 
