+++ CONTRIBUTING.md
@@ -3,12 +3,23 @@
 Thanks for your interest in contributing to BiGSTARS.jl!
 
 Contributions of all kinds are welcome: bug reports, documentation improvements, 
-new examples, performance tweaks, and core features.
+new examples, performance tweaks, and core features. 
+Best way to get in touch is to either just open a GitHub [issue](https://github.com/subhk/BiGSTARS.jl/issues) 
+
+We follow the [ColPrac guide](https://github.com/SciML/ColPrac) for collaborative
+practices. New contributors should make sure to read that guide.
+
+## What can I do?
+ - Bug fixes – if something fails or behaves oddly.
+ - New features / extensions – e.g. new base flows, boundary conditions, parameterizations, or diagnostics.
+ - Performance improvements – more efficient linear algebra, better use of sparse matrices, etc.
+ - Documentation – clarifying docstrings, user docs in docs/, or worked examples in examples/.
+ - Testing – adding coverage for edge cases, new solvers, or new physics.
+
+If you’re interested in working on something, please let us know by commenting on 
+an existing issue or opening a new one. This helps avoid duplicated work and gives 
+us a chance to share any context or guidance that might be useful before you get started.
 
 
-If you’re adding a new feature:
- - Put core logic in src/ (not in examples/).
- - Add tests in test/.
- - If relevant, add or update an example in examples/.
- - Update docs in docs/ or docstrings.
+
 
