### 2025-11-14 subhk/BiGSTARS.jl#14 @alexfikl
Looks good to me! Thanks :grin:
