+++ docs/Project.toml
@@ -1,30 +1,11 @@
 [deps]
-BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
 BiGSTARS = "55e6fa54-20b2-4a8e-ba84-e72db1c430e6"
-CairoMakie = "13f3f980-e62b-5c42-98c6-ff1f3baf88f0"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 DocumenterCitations = "daee34ce-89f3-4625-b898-19384cb65244"
 DocumenterTools = "35a29f4d-8980-5a13-9543-d66fff28ecb8"
-FillArrays = "1a297f60-69ca-5386-bcde-b61e274b549b"
-JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
-LazyGrids = "7031d0ef-c40d-4431-b2f8-61a8d2f650db"
 Literate = "98b081ad-f1c9-55d3-8b20-4c87d4299306"
-ModelingToolkit = "961ee093-0014-501f-94e3-6117800e7a78"
-NonlinearSolve = "8913a72c-1f9b-4ce2-8d82-65094dcecaec"
-Parameters = "d96e819e-fc66-5662-9728-84c9c7592b0a"
-Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
-SparseMatrixDicts = "5cb6c4b0-9b79-11e8-24c9-f9621d252589"
-SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
-StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
-ArnoldiMethod = "ec485272-7323-5ecc-a04f-4719b315124d"
 
 [compat]
-BenchmarkTools = "1.6.0"
-CairoMakie = "0.11, 0.12, 0.13"
 Documenter = "1"
 DocumenterCitations = "1.2"
 Literate = "≥2.9.0"
-Parameters = "0.12.3"
-SpecialFunctions = "2.5.1"
-StaticArrays = "0.12, 1"
-ArnoldiMethod = "0.4.0"
