+++ docs/Project.toml
@@ -5,7 +5,9 @@ DocumenterCitations = "daee34ce-89f3-4625-b898-19384cb65244"
 DocumenterTools = "35a29f4d-8980-5a13-9543-d66fff28ecb8"
 Literate = "98b081ad-f1c9-55d3-8b20-4c87d4299306"
 
+
 [compat]
+BenchmarkTools = "1.6.0"
 Documenter = "1"
 DocumenterCitations = "1.2"
 Literate = "≥2.9.0"
