+++ docs/Project.toml
@@ -4,10 +4,11 @@ Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 DocumenterCitations = "daee34ce-89f3-4625-b898-19384cb65244"
 DocumenterTools = "35a29f4d-8980-5a13-9543-d66fff28ecb8"
 Literate = "98b081ad-f1c9-55d3-8b20-4c87d4299306"
-
+BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
 
 [compat]
 BenchmarkTools = "1.6.0"
 Documenter = "1"
 DocumenterCitations = "1.2"
+DocumenterTools = "1.15.0"
 Literate = "≥2.9.0"
