+++ Project.toml
@@ -7,10 +7,7 @@ version = "2.0.5"
 ArnoldiMethod = "ec485272-7323-5ecc-a04f-4719b315124d"
 Arpack = "7d9fca2a-8960-54d3-9f78-7d1dccf2cb97"
 BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
-Bibliography = "f1be7e48-bf82-45af-a471-ae754a193061"
 Dierckx = "39dd38d3-220a-591b-8e3c-4c3a8c710a94"
-DispatchDoctor = "8d63f2c5-f18a-4cf2-ba9d-b3f60fc568c8"
-DocumenterTools = "35a29f4d-8980-5a13-9543-d66fff28ecb8"
 FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
 FillArrays = "1a297f60-69ca-5386-bcde-b61e274b549b"
 JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
@@ -35,10 +32,7 @@ VectorInterface = "409d34a3-91d5-4945-b6ec-7529ddf182d8"
 ArnoldiMethod = "0.4.0"
 Arpack = "0.5.3"
 BenchmarkTools = "1.5.0"
-Bibliography = "0.3.0"
 Dierckx = "0.5.1"
-DispatchDoctor = "0.4.22"
-DocumenterTools = "0.1.20"
 FFTW = "1"
 FillArrays = "1.13.0"
 JLD2 = "0.4, 0.5"
