+++ docs/Project.toml
@@ -10,5 +10,4 @@ BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
 BenchmarkTools = "1.6.0"
 Documenter = "1"
 DocumenterCitations = "1.2"
-DocumenterTools = "1.15.0"
 Literate = "≥2.9.0"
