+++ Project.toml
@@ -6,6 +6,7 @@ version = "2.0.5"
 [deps]
 ArnoldiMethod = "ec485272-7323-5ecc-a04f-4719b315124d"
 Arpack = "7d9fca2a-8960-54d3-9f78-7d1dccf2cb97"
+BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
 Dierckx = "39dd38d3-220a-591b-8e3c-4c3a8c710a94"
 FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
 FillArrays = "1a297f60-69ca-5386-bcde-b61e274b549b"
@@ -30,6 +31,7 @@ VectorInterface = "409d34a3-91d5-4945-b6ec-7529ddf182d8"
 [compat]
 ArnoldiMethod = "0.4.0"
 Arpack = "0.5.3"
+BenchmarkTools = "1.5.0"
 Dierckx = "0.5.1"
 FFTW = "1"
 FillArrays = "1.13.0"
@@ -51,7 +53,6 @@ VectorInterface = "0.5"
 julia = "1.10, 1.11, 1.12"
 
 [extras]
-BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
