### 2025-11-14 90363c3
Update README with example script execution instructions
Added instructions to run example scripts in Julia.
