+++ README.md
@@ -42,7 +42,7 @@ BiGSTARS.jl requires **Julia 1.6** or newer. Continuous-integration tests curren
 Example scripts can be found in the `examples/` directory. For the clearest overview, we recommend 
 browsing them through the package’s documentation.
 
-You can run the example scripts in the `examples/' directly using the `.toml` files in this directory.
+You can run the example scripts in the `examples/` directly using the `.toml` files in this directory.
 To do so, first start Julia and activate the project for this directory, for example:
 ```
 $ julia --project="path/to/examples"
