+++ README.md
@@ -42,10 +42,10 @@ BiGSTARS.jl requires **Julia 1.6** or newer. Continuous-integration tests curren
 Example scripts can be found in the `examples/` directory. For the clearest overview, we recommend 
 browsing them through the package’s documentation.
 
-You can run the example scripts in the `examples/` directly using the `.toml` files in this directory.
+You can run the example scripts in the `examples/` directly using the `.toml` file.
 To do so, first start Julia and activate the project for this directory, for example:
 ```
-$ julia --project="path/to/examples"
+$ julia --project=.
 ```
 Then, in the Julia REPL, instantiate the project to install the dependencies:
 ```julia
@@ -53,7 +53,7 @@ Then, in the Julia REPL, instantiate the project to install the dependencies:
 ```
 After that, you can run any of the example scripts with:
 ```julia
-include("path/to/examples/example_script.jl")
+include("examples/example_script.jl")
 ```
 
 
