### 2025-11-14 225e8a6
Update README for clarity on project activation
