+++ docs/Project.toml
@@ -6,6 +6,7 @@ DocumenterTools = "35a29f4d-8980-5a13-9543-d66fff28ecb8"
 Literate = "98b081ad-f1c9-55d3-8b20-4c87d4299306"
 BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
 CairoMakie="13f3f980-e62b-5c42-98c6-ff1f3baf88f0"
+LazyGrids = "7031d0ef-c40d-4431-b2f8-61a8d2f650db"
 
 [compat]
 BenchmarkTools = "1.6.0"
