+++ examples/Eady.jl
@@ -116,15 +116,8 @@
 # ## Load required packages
 using LazyGrids
 using LinearAlgebra
-using Printf
-using StaticArrays
 using SparseArrays
-using SparseMatrixDicts
 using FillArrays
-using SpecialFunctions
-using Parameters
-using Test
-using BenchmarkTools
 using JLD2
 using Parameters: @with_kw
 
