+++ docs/Project.toml
@@ -5,11 +5,16 @@ DocumenterCitations = "daee34ce-89f3-4625-b898-19384cb65244"
 DocumenterTools = "35a29f4d-8980-5a13-9543-d66fff28ecb8"
 Literate = "98b081ad-f1c9-55d3-8b20-4c87d4299306"
 BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
-CairoMakie="13f3f980-e62b-5c42-98c6-ff1f3baf88f0"
 LazyGrids = "7031d0ef-c40d-4431-b2f8-61a8d2f650db"
+FillArrays = "1a297f60-69ca-5386-bcde-b61e274b549b"
+LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
+SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
+JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
+Parameters = "d96e819e-fc66-5662-9728-84c9c7592b0a"
 
 [compat]
 BenchmarkTools = "1.6.0"
 Documenter = "1"
 DocumenterCitations = "1.2"
 Literate = "≥2.9.0"
+
