+++ Project.toml
@@ -34,7 +34,6 @@ BenchmarkTools = "1.5.0"
 Dierckx = "0.5.1"
 FFTW = "1"
 FillArrays = "1.13.0"
-JLD2 = "0.4, 0.5"
 KrylovKit = "0.9.5"
 LazyGrids = "1.1.0"
 LinearAlgebra = "1.6"
