+++ examples/Eady.jl
@@ -118,6 +118,7 @@ using LazyGrids
 using LinearAlgebra
 using SparseArrays
 using FillArrays
+using Printf
 using JLD2
 using Parameters: @with_kw
 
