+++ docs/Project.toml
@@ -10,7 +10,6 @@ FillArrays = "1a297f60-69ca-5386-bcde-b61e274b549b"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
-Parameters = "d96e819e-fc66-5662-9728-84c9c7592b0a"
 
 [compat]
 BenchmarkTools = "1.6.0"
