+++ examples/Eady.jl
@@ -119,7 +119,6 @@ using LinearAlgebra
 using SparseArrays
 using FillArrays
 using Printf
-using JLD2
 using Parameters: @with_kw
 
 using BiGSTARS
