+++ examples/Stone1971.jl
@@ -240,7 +240,6 @@ using LinearAlgebra
 using Printf
 using SparseArrays
 using FillArrays
-using JLD2
 using Parameters: @with_kw
 
 using BiGSTARS
