### 2025-11-14 de11f03
Add JLD2 import to example scripts
Added 'using JLD2' to Eady.jl, Stone1971.jl, and rRBC.jl to enable JLD2 functionality in these example files.
