+++ examples/rRBC.jl
@@ -172,6 +172,7 @@ using LinearAlgebra
 using Printf
 using StaticArrays
 using FillArrays
+using JLD2
 using Parameters: @with_kw
 
 using BiGSTARS
