### 2025-11-14 15c5081
Replace StaticArrays with SparseArrays in rRBC.jl
Updated the import in examples/rRBC.jl to use SparseArrays instead of StaticArrays, likely to support sparse matrix operations in the example.
