### 2025-11-14 subhk/BiGSTARS.jl#15 @subhk
@alexfikl Thanks a lot for the suggestions!
I’ve now added a CONTRIBUTING.md with guidelines for code development, and also documented how to run the scripts in examples/ using the local project environment.
