### 2025-11-15 96f0238
CompatHelper: bump compat for JLD2 to 0.6, (keep existing compat)
