### 2025-11-15 1bfd476
Update CONTRIBUTING.md for clarity
Removed 'parameterizations' from new features/extensions section.
