### 2025-11-22 e5c4f24
Update CONTRIBUTING.md
