### 2025-11-22 subhk/BiGSTARS.jl#15 @subhk
@alexfikl  Sorry for my delayed response.
Thank you for the reference!

Now, I have modified the contributing part here
https://github.com/subhk/BiGSTARS.jl?tab=contributing-ov-file

I included details on setting up the development environment, running tests, and building documentation.

Please let me know your thoughts.

P.S.: I really appreciate your contributions to improve this project!  :grin:
