### 2025-11-22 subhk/BiGSTARS.jl#15 @alexfikl
> I included details on setting up the development environment, running tests, and building documentation.

That's lovely now! Thank you :grin:
