### 2025-11-22 #9368 @alexfikl
@jatkinson1000 I have finished my review! The software side is well-written and well-documented, with several examples to showcase what it can do (that I managed to get running pretty quickly). The manuscript itself also contains all the JOSS-recommended sections and clearly describes what niche BiGSTARS.jl is trying to fill. All in all, I'm recommending this for publication :grin:

The only worry I have is that the project seems to have been started in July 2025 and hasn't seen much development after the initial push. I'm not familiar enough with JOSS guidelines about that, but it's not a deal breaker for me.
