### 2025-11-24 subhk/BiGSTARS.jl#20 @ranocha
Julia project requirements
Xref https://github.com/openjournals/joss-reviews/issues/9368

https://github.com/subhk/BiGSTARS.jl/blob/4dbe35dc882b2757b608b0543d56249f8891bf09/Project.toml#L53

This is just the same as

```toml
julia = "1.10"
```

Moreover, it contradicts

https://github.com/subhk/BiGSTARS.jl/blob/4dbe35dc882b2757b608b0543d56249f8891bf09/README.md?plain=1#L37

and the docs (https://subhk.github.io/BiGSTARSDocumentation/stable/installation_instructions/)
