### 2025-11-24 subhk/BiGSTARS.jl#22 @ranocha
Links to related packages
Xref https://github.com/openjournals/joss-reviews/issues/9368

It would be helpful if there could be links to related packages, e.g., the eigensolvers mentioned on https://subhk.github.io/BiGSTARSDocumentation/stable/method/
