### 2025-11-24 subhk/BiGSTARS.jl#23 @ranocha
Why `myfindall`
Xref https://github.com/openjournals/joss-reviews/issues/9368

Is was surprised to see https://subhk.github.io/BiGSTARSDocumentation/stable/basic_functionality/#myfindall(condition,-x). Why is `myfindall` required? Can't you just use `findall` from Base?
