### 2025-11-24 subhk/BiGSTARS.jl#26 @ranocha
Statement of need in the documentation
Xref https://github.com/openjournals/joss-reviews/issues/9368

While the statement of need is explained in the paper, it does not become evident to me from the README and the documentation.
