### 2025-11-28 0b99f10
CompatHelper: bump compat for KrylovKit to 0.10, (keep existing compat)
