### 2025-11-29 a961b9e
renaming ARPACK to Arpack
