+++ paper/paper.bib
@@ -17,21 +17,24 @@
   number={1},
   pages={319--352},
   year={2011},
-  publisher={Annual Reviews}
+  publisher={Annual Reviews},
+  doi={10.1146/annurev-fluid-122109-160801}
 }
 
 @book{vallis2017atmospheric,
   title={Atmospheric and oceanic fluid dynamics},
   author={Vallis, Geoffrey K},
   year={2017},
-  publisher={Cambridge University Press}
+  publisher={Cambridge University Press},
+  doi={10.1017/9781107588417}
 }
 
 @book{pedlosky2013geophysical,
   title={Geophysical fluid dynamics},
   author={Pedlosky, Joseph},
   year={2013},
-  publisher={Springer Science \& Business Media}
+  publisher={Springer Science \& Business Media},
+  doi={10.1007/978-1-4614-7572-9}
 }
 
 
@@ -61,7 +64,8 @@ year = {2021}}
   number={2189},
   pages={20160117},
   year={2016},
-  publisher={The Royal Society Publishing}
+  publisher={The Royal Society Publishing},
+  doi={10.1098/rspa.2016.0117}
 }
 
 
@@ -102,7 +106,7 @@ url = {https://github.com/JuliaLinearAlgebra/ArnoldiMethod.jl}
   title={Hydrodynamic stability},
   author={Drazin, Philip G and Reid, William Hill},
   year={2004},
-  publisher={Cambridge university press}
+  publisher={Cambridge university press},
+  doi={10.1017/CBO9780511809068}
 }
-
 
