### 2025-11-30 8d0ef57
added statement of need in README and docs
