+++ README.md
@@ -3,6 +3,8 @@
 <!-- description --> 
   **Bi**-**G**lobal **St**ability **A**nalysis of **R**otating **S**tratified Flows (BiGSTARS ⭐): A linear stability analysis tool for Geophysical flows with Julia. 
 
+Bi-global stability analysis sits between 1D (too idealized) and fully tri-global (often too expensive) approaches. BiGSTARS.jl gives geophysical fluid dynamicists a practical middle ground: spectral Chebyshev–Fourier discretizations, shift-and-invert eigen solvers, and ready-to-run benchmarks for rotating, stratified flows so you can resolve key instabilities without massive HPC resources.
+
  <!-- Badges -->
  <p align="left">
     <a href="https://github.com/subhk/BiGSTARS.jl/actions/workflows/CI.yml">
