+++ docs/src/index.md
@@ -4,6 +4,8 @@
 
 `BiGSTARS.jl` — a Julia toolkit for bi-global linear-stability analysis of geophysical flows, built on a Chebyshev–Fourier spectral-collocation method.
 
+Bi-global analysis bridges the gap between 1D stability tools (often too idealized) and fully 3D tri-global solvers (often too costly). BiGSTARS.jl offers a practical middle ground for rotating, stratified flows: spectral discretizations, shift-and-invert eigen solvers, and curated benchmarks so you can capture key instabilities without heavy HPC requirements.
+
 ## Examples
 
 The examples illustrate the key features of every module.  
