### 2025-12-01 8590a75
modified Summary in paper.md
