### 2025-12-01 8fce558
Update paper.md
