### 2025-12-01 5d5d96f
fixed memory allocations
