+++ test/Stone1971.jl
@@ -71,7 +71,6 @@ function generalized_EigValProb(prob, grid, params)
     s₂ = size(I⁰, 2);
 
     ## the horizontal Laplacian operator
-    ∇ₕ² = SparseMatrixCSC(Zeros(N, N))
     ∇ₕ² = (1.0 * prob.D²ʸ - 1.0 * params.k^2 * I⁰)
 
     ## inverse of the horizontal Laplacian operator
