### 2025-12-01 35793dc
provide details on Chebyshev method
