### 2025-12-01 706c81e
Update paper.bib
