+++ paper/paper.bib
@@ -110,3 +110,10 @@ url = {https://github.com/JuliaLinearAlgebra/ArnoldiMethod.jl}
   doi={10.1017/CBO9780511809068}
 }
 
+@book{trefethen2000spectral,
+  title={Spectral methods in MATLAB},
+  author={Trefethen, Lloyd N},
+  year={2000},
+  publisher={SIAM}
+}
+
