### 2025-12-01 d3c2f89
Update Chebyshev.jl
