+++ src/construct_linear_map.jl
@@ -10,7 +10,14 @@ function (M::ShiftAndInvert)(y, x)
     ldiv!(y, M.A_lu, M.temp)
 end
 
+# Default construction (kept for compatibility)
 function construct_linear_map(A, B)
-    ShiftAndInvert(factorize(A), B, Vector{eltype(A)}(undef, size(A,1))) |>
-    M -> LinearMap{eltype(A)}(M, size(A,1), ismutating=true)
+    temp = Vector{eltype(A)}(undef, size(A, 1))
+    construct_linear_map(factorize(A), B, temp)
+end
+
+# New overload that reuses an existing factorization and workspace vector.
+function construct_linear_map(A_fact, B, temp::AbstractVector)
+    ShiftAndInvert(A_fact, B, temp) |>
+    M -> LinearMap{eltype(temp)}(M, length(temp), ismutating=true)
 end
