+++ src/gradient.jl
@@ -1,3 +1,5 @@
+import Dierckx
+
 """
     GradientCalculator{T<:AbstractFloat}
 
@@ -90,13 +92,14 @@ function (calc::GradientCalculator{S})(grad::AbstractArray{T},
     @assert nd ≤ 3 "GradientCalculator currently only supports 1D, 2D, or 3D arrays"
 
     n = size(f)
+    ws = _spline_workspace(length(x), calc.spline_degree)
 
     if nd == 1
-        _compute_gradient_1d!(calc, grad, f, x)
+        _compute_gradient_1d!(calc, grad, f, x, ws)
     elseif nd == 2
-        _compute_gradient_2d!(calc, grad, f, x, dims, n)
+        _compute_gradient_2d!(calc, grad, f, x, dims, n, ws)
     else # nd == 3
-        _compute_gradient_3d!(calc, grad, f, x, dims, n)
+        _compute_gradient_3d!(calc, grad, f, x, dims, n, ws)
     end
 
     return grad
@@ -161,13 +164,98 @@ Compute 1st derivative using extrapolation boundary conditions.
 """
 extrapolate_gradient(f, x; dims=1) = GradientCalculator("extrapolate", 3, 1)(f, x; dims=dims)
 
+# ────────────────────────────────────────────────────────────────────────────────
+# Lightweight spline workspace to reuse allocations across slices
+# ────────────────────────────────────────────────────────────────────────────────
+mutable struct Spline1DWorkspace
+    x::Vector{Float64}
+    y::Vector{Float64}
+    w::Vector{Float64}
+    t::Vector{Float64}
+    c::Vector{Float64}
+    wrk::Vector{Float64}
+    iwrk::Vector{Int32}
+    k::Int
+end
+
+function _spline_workspace(m::Int, k::Int)
+    nest = max(m + k + 1, 2k + 3)
+    lwrk = m * (k + 1) + nest * (7 + 3k)
+    return Spline1DWorkspace(
+        Vector{Float64}(undef, m),          # x buffer
+        Vector{Float64}(undef, m),          # y buffer
+        ones(Float64, m),                   # weights (all ones)
+        Vector{Float64}(undef, nest),       # knots
+        Vector{Float64}(undef, nest),       # coeffs (upper bound)
+        Vector{Float64}(undef, lwrk),       # work array
+        Vector{Int32}(undef, nest),         # integer workspace
+        k
+    )
+end
+
+@inline function _ensure_workspace!(ws::Spline1DWorkspace, m::Int, k::Int)
+    ws.k = k
+    if length(ws.x) != m
+        resize!(ws.x, m)
+        resize!(ws.y, m)
+        resize!(ws.w, m)
+        fill!(ws.w, 1.0)
+    end
+
+    nest = max(m + k + 1, 2k + 3)
+    lwrk = m * (k + 1) + nest * (7 + 3k)
+
+    length(ws.t) < nest && resize!(ws.t, nest)
+    length(ws.c) < nest && resize!(ws.c, nest)
+    length(ws.wrk) < lwrk && resize!(ws.wrk, lwrk)
+    length(ws.iwrk) < nest && resize!(ws.iwrk, nest)
+
+    return nest, lwrk
+end
+
+function _fit_spline!(ws::Spline1DWorkspace,
+                      x::AbstractVector,
+                      y::AbstractVector,
+                      k::Int,
+                      bc::AbstractString)
+    m = length(x)
+    nest, lwrk = _ensure_workspace!(ws, m, k)
+
+    copyto!(ws.x, x)
+    copyto!(ws.y, y)
+    fill!(ws.w, 1.0)
+
[…]
