### 2025-12-06 0a6f09b
Update paper.md
