+++ examples/rRBC.jl
@@ -168,7 +168,6 @@ using LazyGrids
 using LinearAlgebra
 using Printf
 using SparseArrays
-using FillArrays
 using JLD2
 using Parameters: @with_kw
 
