+++ examples/visualization.jl
@@ -2,14 +2,8 @@
 using LazyGrids
 using LinearAlgebra
 using Printf
-using StaticArrays
 using SparseArrays
-using SparseMatrixDicts
-using FillArrays
-using SpecialFunctions
 using Parameters
-using Test
-using BenchmarkTools
 using JLD2
 using Dierckx
 using Parameters: @with_kw
