### 2025-12-06 2d0a85e
Update paper.bib
