+++ paper/paper.bib
@@ -30,19 +30,22 @@
 }
 
 @book{pedlosky2013geophysical,
-  title={Geophysical fluid dynamics},
-  author={Pedlosky, Joseph},
-  year={2013},
-  publisher={Springer Science \& Business Media},
-  doi={10.1007/978-1-4614-7572-9}
+  title     = {Geophysical Fluid Dynamics},
+  author    = {Pedlosky, Joseph},
+  year      = {1987},
+  edition   = {2nd},
+  publisher = {Springer},
+  doi       = {10.1007/978-1-4612-4650-3}
 }
 
 
-@article{abernathey2022pyqg,
-  title={pyqg/pyqg: v0. 7.2},
-  author={Abernathey, Ryan and Rocha, CB and Ross, A and Jansen, M and Li, Z and Poulin, FJ},
-  journal={Zenodo},
-  year={2022}
+@software{abernathey2022pyqg,
+  author    = {Abernathey, Ryan and Rocha, Cesar B and Ross, Andrew and Jansen, Malte and Li, Ziwei and Poulin, Francis J},
+  title     = {pyqg: Python Quasigeostrophic Model},
+  year      = {2022},
+  publisher = {Zenodo},
+  version   = {v0.7.2},
+  doi       = {10.5281/zenodo.7348988}
 }
 
 @article{eigentools_2021,
@@ -87,9 +90,10 @@ year = {2021}}
 }
 
 @software{Stoppels,
-author = {Stoppels, Harmen Taeke and Nyman, Lauri},
-title = {ArnoldiMethod.jl: Arnoldi Method with Krylov-Schur restart natively in Julia},
-url = {https://github.com/JuliaLinearAlgebra/ArnoldiMethod.jl}
+  author  = {Stoppels, Harmen Taeke and Nyman, Lauri},
+  title   = {ArnoldiMethod.jl: Arnoldi Method with Krylov-Schur restart},
+  year    = {2024},
+  url     = {https://github.com/JuliaLinearAlgebra/ArnoldiMethod.jl}
 }
 
 
@@ -111,9 +115,10 @@ url = {https://github.com/JuliaLinearAlgebra/ArnoldiMethod.jl}
 }
 
 @book{trefethen2000spectral,
-  title={Spectral methods in MATLAB},
-  author={Trefethen, Lloyd N},
-  year={2000},
-  publisher={SIAM}
+  title     = {Spectral Methods in MATLAB},
+  author    = {Trefethen, Lloyd N},
+  year      = {2000},
+  publisher = {SIAM},
+  doi       = {10.1137/1.9780898719598}
 }
 
