### 2025-12-06 64fc489
Update paper.bib
