+++ paper/paper.bib
@@ -45,7 +45,7 @@
   year      = {2022},
   publisher = {Zenodo},
   version   = {v0.7.2},
-  doi       = {10.5281/zenodo.7348988}
+  doi       = {10.5281/zenodo.6563667}
 }
 
 @article{eigentools_2021,
