### 2025-12-06 ef5afa1
Update paper.bib
