+++ paper/paper.bib
@@ -18,7 +18,7 @@
   pages={319--352},
   year={2011},
   publisher={Annual Reviews},
-  doi={10.1146/annurev-fluid-122109-160801}
+  doi={10.1146/annurev-fluid-122109-160705}
 }
 
 @book{vallis2017atmospheric,
@@ -26,7 +26,7 @@
   author={Vallis, Geoffrey K},
   year={2017},
   publisher={Cambridge University Press},
-  doi={10.1017/9781107588417}
+  doi={10.2277/0521849691}
 }
 
 @book{pedlosky2013geophysical,
@@ -35,7 +35,7 @@
   year      = {1987},
   edition   = {2nd},
   publisher = {Springer},
-  doi       = {10.1007/978-1-4612-4650-3}
+  doi       = {10.1007/978-3-662-25730-2}
 }
 
 
@@ -79,8 +79,8 @@ year = {2021}}
   year         = 2025,
   publisher    = {Zenodo},
   version      = {v0.10.0},
-  doi          = {10.5281/zenodo.16710486},
-  url          = {https://doi.org/10.5281/zenodo.16710486},
+  doi          = {10.5281/zenodo.17316953},
+  url          = {https://doi.org/10.5281/zenodo.17316953},
   swhid        = {swh:1:dir:fbc17c543b3f2dd691adaf31c1e2045a87c60337
                    ;origin=https://doi.org/10.5281/zenodo.10622234;vi
                    sit=swh:1:snp:7ea589674b018280926ea7a2ac2fa246ae3a
@@ -111,7 +111,7 @@ year = {2021}}
   author={Drazin, Philip G and Reid, William Hill},
   year={2004},
   publisher={Cambridge university press},
-  doi={10.1017/CBO9780511809068}
+  doi={10.1017/CBO9780511616938}
 }
 
 @book{trefethen2000spectral,
