### 2025-12-06 374165c
Update paper.bib
