+++ paper/paper.bib
@@ -26,7 +26,7 @@
   author={Vallis, Geoffrey K},
   year={2017},
   publisher={Cambridge University Press},
-  doi={10.2277/0521849691}
+  doi={10.1017/9781107588417}
 }
 
 @book{pedlosky2013geophysical,
