+++ Project.toml
@@ -50,7 +50,7 @@ ToeplitzMatrices = "0.8.4"
 Trapz = "2.0.3"
 Tullio = "0.3.6"
 VectorInterface = "0.5"
-julia = "1.10, 1.11, 1.12"
+julia = "1.10"
 
 [extras]
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
