### 2025-12-06 c39f3f6
Update installation_instructions.md
