+++ docs/src/installation_instructions.md
@@ -24,10 +24,5 @@ We're more than happy to help you get your model up and running again.
 
 
 !!! warn "Julia 1.10 or newer required"
-    The latest version of `BiGSTARS.jl` requires **at least Julia v1.6** to run.  
-    Installing `BiGSTARS.jl` with an older version of Julia will instead install the latest version that is compatible with your Julia installation.
-
-    BiGSTARS.jl is continuously tested on **Julia v1.10** (the current long-term support), **v1.11** and 
-    **v1.12**.  
-    _For optimal stability and reproducibility, we strongly recommend using one of these tested Julia versions._
+    BiGSTARS.jl is continuously tested on Julia v1.10 (the current long-term support), v1.11 and v1.12. For optimal stability and reproducibility, we strongly recommend using one of these tested Julia versions.
 
