### 2025-12-06 subhk/BiGSTARS.jl#20 @subhk
Thank you for catching this inconsistency. You are correct that julia = "1.10, 1.11, 1.12" is equivalent to julia = "1.10" in Julia's Pkg compat specification, and the README/docs incorrectly stated Julia 1.6 as the minimum version.

We have fixed this by:
1. Make Project.toml to julia = "1.10"
2. Updating README.md to state "BiGSTARS.jl requires Julia 1.10 or newer"
3. Correcting the relevant section of documentation to be consistent
