+++ test/Stone1971.jl
@@ -10,14 +10,9 @@
 using LazyGrids
 using LinearAlgebra
 using Printf
-using StaticArrays
 using SparseArrays
-using SparseMatrixDicts
-using FillArrays
-using SpecialFunctions
 using Parameters
 using Test
-using BenchmarkTools
 using JLD2
 using BiGSTARS
 using BiGSTARS: AbstractParams
