+++ examples/Stone1971.jl
@@ -239,7 +239,6 @@ using LazyGrids
 using LinearAlgebra
 using Printf
 using SparseArrays
-using FillArrays
 using JLD2
 using Parameters: @with_kw
 
