+++ Project.toml
@@ -6,10 +6,8 @@ version = "2.0.5"
 [deps]
 ArnoldiMethod = "ec485272-7323-5ecc-a04f-4719b315124d"
 Arpack = "7d9fca2a-8960-54d3-9f78-7d1dccf2cb97"
-BenchmarkTools = "6e4b80f9-dd63-53aa-95a3-0cdb28fa8baf"
 Dierckx = "39dd38d3-220a-591b-8e3c-4c3a8c710a94"
 FFTW = "7a1cc6ca-52ef-59f5-83cd-3a7055c09341"
-FillArrays = "1a297f60-69ca-5386-bcde-b61e274b549b"
 JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
 KrylovKit = "0b1a1467-8014-51b9-945f-bf0ae24f4b77"
 LazyGrids = "7031d0ef-c40d-4431-b2f8-61a8d2f650db"
@@ -17,38 +15,30 @@ LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 LinearMaps = "7a12625a-238d-50fd-b39a-03d52299707e"
 Parameters = "d96e819e-fc66-5662-9728-84c9c7592b0a"
 Printf = "de0858da-6303-5e67-8744-51eddeeeb8d7"
-Reexport = "189a3867-3050-52da-a836-e630ba90ab69"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 SparseMatrixDicts = "5cb6c4b0-9b79-11e8-24c9-f9621d252589"
 SpecialFunctions = "276daf66-3868-5448-9aa4-cd146d93841b"
 StaticArrays = "90137ffa-7385-5640-81b9-e52037218182"
 Statistics = "10745b16-79ce-11e8-11f9-7d13ad32a3b2"
 ToeplitzMatrices = "c751599d-da0a-543b-9d20-d0a503d91d24"
-Trapz = "592b5752-818d-11e9-1e9a-2b8ca4a44cd1"
-Tullio = "bc48ee85-29a4-5162-ae0b-a64e1601d4bc"
 VectorInterface = "409d34a3-91d5-4945-b6ec-7529ddf182d8"
 
 [compat]
 ArnoldiMethod = "0.4.0"
 Arpack = "0.5.3"
-BenchmarkTools = "1.5.0"
 Dierckx = "0.5.1"
 FFTW = "1"
-FillArrays = "1.13.0"
 JLD2 = "0.4, 0.5"
 KrylovKit = "0.9.5"
 LazyGrids = "1.1.0"
 LinearAlgebra = "1.6"
 LinearMaps = "3.11.3"
 Parameters = "0.12.1"
-Reexport = "0.2, 1"
 SparseMatrixDicts = "0.3.0"
 SpecialFunctions = "2.5.1"
 StaticArrays = "0.12, 1"
 Statistics = "1.6"
 ToeplitzMatrices = "0.8.4"
-Trapz = "2.0.3"
-Tullio = "0.3.6"
 VectorInterface = "0.5"
 julia = "1.10"
 
