+++ docs/Project.toml
@@ -2,10 +2,8 @@
 BiGSTARS = "55e6fa54-20b2-4a8e-ba84-e72db1c430e6"
 Documenter = "e30172f5-a6a5-5a46-863b-614d45cd2de4"
 DocumenterCitations = "daee34ce-89f3-4625-b898-19384cb65244"
-DocumenterTools = "35a29f4d-8980-5a13-9543-d66fff28ecb8"
 Literate = "98b081ad-f1c9-55d3-8b20-4c87d4299306"
 LazyGrids = "7031d0ef-c40d-4431-b2f8-61a8d2f650db"
-FillArrays = "1a297f60-69ca-5386-bcde-b61e274b549b"
 LinearAlgebra = "37e2e46d-f89d-539d-b4ee-838fcccc9c8e"
 SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
