+++ Project.toml
@@ -37,7 +37,9 @@ VectorInterface = "0.5"
 julia = "1.10"
 
 [extras]
+JLD2 = "033835bb-8acc-5ee8-8aae-3f567f8a3819"
+LazyGrids = "7031d0ef-c40d-4431-b2f8-61a8d2f650db"
 Test = "8dfed614-e22c-5e08-85e1-65c5234f0b40"
 
 [targets]
-test = ["Test"]
+test = ["JLD2", "LazyGrids", "Test"]
