### 2025-12-13 78b5beb
CompatHelper: bump compat for KrylovKit to 0.10, (keep existing compat)
