+++ Project.toml
@@ -1,7 +1,7 @@
 name = "BiGSTARS"
 uuid = "55e6fa54-20b2-4a8e-ba84-e72db1c430e6"
-authors = ["Subhajit Kar <subhajitkar19@gmail.com>"]
 version = "2.0.6"
+authors = ["Subhajit Kar <subhajitkar19@gmail.com>"]
 
 [deps]
 ArnoldiMethod = "ec485272-7323-5ecc-a04f-4719b315124d"
@@ -25,7 +25,7 @@ ArnoldiMethod = "0.4.0"
 Arpack = "0.5.3"
 Dierckx = "0.5.1"
 FFTW = "1"
-KrylovKit = "0.9.5"
+KrylovKit = "0.9.5, 0.10"
 LinearAlgebra = "1.6"
 LinearMaps = "3.11.3"
 Parameters = "0.12.1"
