### 2025-12-18 d9e95b3
Update paper.md
Added acknowledgements for funding support and discussions.
