+++ paper/paper.md
@@ -23,7 +23,7 @@ bibliography: paper.bib
 `BiGSTARS.jl` is a software package written in the Julia programming language [@Bezanson2017] for investigating the stability of fluid flows relevant to the atmosphere and ocean. It performs linear stability analysis, a technique that determines whether small disturbances introduced into a flow will grow or decay over time. Unlike simpler one-dimensional approaches, bi-global analysis can handle flows whose properties vary in two spatial directions, enabling the study of realistic features such as fronts, jets, and vortices. The package is specifically designed for rotating and stratified fluids, capturing the physical processes essential to geophysical fluid dynamics.
 
 # Statement of need
-Linear stability analysis investigates the growth or decay of small perturbations about a basic state by linearizing the governing equations and solving the resulting eigenvalue problem [@drazin2004]. In geophysical fluid dynamics, the interplay between rotation (the Coriolis force) and stratification (the buoyancy force) gives rise to multiple instability mechanisms, including baroclinic instability driven by vertical shear and barotropic instability driven by horizontal shear [@pedlosky2013geophysical]. Such analyses are fundamental for understanding the onset of turbulence, the formation of eddies, and the associated energy transfer across scales in both oceanic and atmospheric systems.
+Linear stability analysis investigates the growth or decay of small perturbations about a basic state by linearizing the governing equations and solving the resulting eigenvalue problem [@drazin2004]. In geophysical fluid dynamics, the interplay between rotation (the Coriolis force) and stratification (the buoyancy force) gives rise to multiple instability mechanisms, including baroclinic instability driven by buoyancy flux and barotropic instability driven by horizontal shear [@pedlosky2013geophysical]. Such analyses are fundamental for understanding the onset of turbulence, the formation of eddies, and the associated energy transfer across scales in both oceanic and atmospheric systems.
 
 The complexity and dimensionality of stability analyses have evolved substantially, giving rise to distinct methodological approaches with varying computational requirements [@theofilis2011global]. The classical one-dimensional (1D) stability analysis assumes that the basic state varies along only a single spatial direction. While computationally efficient, this approach presumes spatial homogeneity in the remaining directions and therefore often fails to capture the full dynamics of realistic geophysical flows. At the opposite extreme, the tri-global (3D global) stability analysis allows variations in all three spatial directions, providing the most complete representation of the underlying physics. However, this generality requires substantial computational resources, often beyond the reach of standard research computing infrastructures.
 
@@ -79,6 +79,6 @@ growthrate and imaginary part denotes the frequency. The variables with superscr
 
 # Acknowledgements
 
-The author gratefully acknowledges Sohan Suresan for the stimulating discussions.
+The author gratefully acknowledges Sohan Suresan for the stimulating discussions. The author acknowledges support from the Israel Science Foundation (ISF; grant 2054/23) and from the European Union through the European Research Council (ERC, 401 ML Transport 101163887) project. Views and opinions expressed are, however, those of the author(s) only and do not necessarily reflect those of the European Union or the European Research Council. Neither the European Union nor the granting authority can be held responsible for them.
 
 # References
