### 2026-01-23 #9368 @jatkinson1000
Hi @subhk I have made some very minor editorial suggestions in a pull request at https://github.com/subhk/BiGSTARS.jl/pull/32
Thank you for a nice, well-written paper, it was an interesting read.

I have checked the references and repository, so if you could please proceed with the other items on the checklist above (archive, version etc.) and let me know when you are ready that would be great, thank you.
