### 2026-01-29 #9368 @subhk
Hi @jatkinson1000,
                                                                                                                                    Thank you — I'm glad you found it interesting!
                                                                                                                                      I've also completed the archive and version items:

 - Archive DOI: 10.5281/zenodo.18385010
 - Version: 2.0.8
