### 2026-02-09 #9368 @jatkinson1000
@subhk sorry to be pedantic but the paper and archive title and authors need to match exactly for JOSS.
Could you add ", Israel" to your zenodo affiliation please?
