### 2026-02-09 #9368 @jatkinson1000
Thanks @subhk this all looks good to me!
I'm recommending acceptance and publication and this will now be passed up to the track editor for final checks and release.

Thanks also to our dedicated volunteer reviewers @ranocha and @alexfikl
I hope you enjoyed the process, and look forward to seeing you again in a future review.
As a reminder, you can [register as a JOSS reviewer](https://reviewers.joss.theoj.org/join) if you haven't already.

Thanks for your submission and participation in the open review process.
