### 2026-02-09 #9368 @editorialbot
:wave: @openjournals/ese-eics, this paper is ready to be accepted and published.
