### 2026-02-13 #9368 @kthyng
@subhk Please check the capitalization in your references. You can preserve capitalization by placing {} around characters/words in your .bib file.
