### 2026-02-13 3fc7c7c
Update paper.bib
