+++ paper/paper.bib
@@ -10,7 +10,7 @@
 }
 
 @article{theofilis2011global,
-  title={Global linear instability},
+  title={{Global Linear Instability}},
   author={Theofilis, Vassilios},
   journal={Annual Review of Fluid Mechanics},
   volume={43},
@@ -22,7 +22,7 @@
 }
 
 @book{vallis2017atmospheric,
-  title={Atmospheric and oceanic fluid dynamics},
+  title={{Atmospheric and Oceanic Fluid Dynamics}},
   author={Vallis, Geoffrey K},
   year={2017},
   publisher={Cambridge University Press},
@@ -30,7 +30,7 @@
 }
 
 @book{pedlosky2013geophysical,
-  title     = {Geophysical Fluid Dynamics},
+  title     = {{Geophysical Fluid Dynamics}},
   author    = {Pedlosky, Joseph},
   year      = {1987},
   edition   = {2nd},
