### 2026-02-13 7fa6ffc
Update paper.bib
