+++ paper/paper.bib
@@ -59,7 +59,7 @@ volume = {6},
 year = {2021}}
 
 @article{mcwilliams2016,
-  title={Submesoscale currents in the ocean},
+  title={{Submesoscale currents in the ocean}},
   author={McWilliams, James C},
   journal={Proceedings of the Royal Society A: Mathematical, Physical and Engineering Sciences},
   volume={472},
@@ -89,7 +89,7 @@ year = {2021}}
 
 @software{Stoppels,
   author  = {Stoppels, Harmen Taeke and Nyman, Lauri},
-  title   = {ArnoldiMethod.jl: Arnoldi Method with Krylov-Schur restart},
+  title   = {{ArnoldiMethod.jl: Arnoldi Method with Krylov-Schur restart}},
   year    = {2024},
   url     = {https://github.com/JuliaLinearAlgebra/ArnoldiMethod.jl}
 }
@@ -104,7 +104,7 @@ year = {2021}}
 }
 
 @book{drazin2004,
-  title={Hydrodynamic stability},
+  title={{Hydrodynamic stability}},
   author={Drazin, Philip G and Reid, William Hill},
   year={2004},
   publisher={Cambridge university press},
@@ -112,7 +112,7 @@ year = {2021}}
 }
 
 @book{trefethen2000spectral,
-  title     = {Spectral Methods in MATLAB},
+  title     = {{Spectral Methods in MATLAB}},
   author    = {Trefethen, Lloyd N},
   year      = {2000},
   publisher = {SIAM},
