### 2026-02-13 #9368 @subhk
@kthyng Thanks for the note. I have fixed the capitalization issues in the references and used braces in the .bib entries where needed to preserve capitalization.
